+++ src/main/java/com/kb04/starroad/Controller/PolicyController.java
@@ -1,18 +1,13 @@
 package com.kb04.starroad.Controller;
 
-import com.kb04.starroad.Dto.PolicyDto;
-import com.kb04.starroad.Dto.ProductDto;
+import com.kb04.starroad.Dto.policy.PolicyDto;
 import com.kb04.starroad.Repository.PolicyRepository;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.annotation.PostConstruct;
 import java.util.List;
 
 @RestController
