+++ src/main/java/com/kb04/starroad/Dto/BaseRateDto.java
@@ -1,36 +1,21 @@
 package com.kb04.starroad.Dto;
 
+import com.kb04.starroad.Dto.product.ProductDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
-
 @Data
-@Entity
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-@Table(name = "base_rate")
-@SequenceGenerator(name = "base_rate_seq", sequenceName = "base_rate_seq", allocationSize = 1)
 public class BaseRateDto {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "base_rate_seq")
     private int no;
-
-    @Column(name = "min_period", nullable = false)
     private int minPeriod;
-
-    @Column(name = "max_period", nullable = false)
     private int maxPeriod;
-
-    @Column(nullable = false)
     private double rate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "prod_no", nullable = false)
     private ProductDto productDto;
 
 }
