--- src/main/java/com/kb04/starroad/Dto/CommentDto.java
@@ -1,38 +1,0 @@
-package com.kb04.starroad.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-@Table(name = "comments")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class CommentDto {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "comment_seq")
-    @SequenceGenerator(name = "comment_seq", sequenceName = "COMMENT_SEQ", allocationSize = 1)
-    @Column(name = "no")
-    private int no;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "board_no")
-    private BoardDto board;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_no")
-    private MemberDto member;
-
-    @Column(name = "regdate", nullable = false)
-    private Date regdate;
-
-    @Column(name = "content", nullable = false, length = 2000)
-    private String content;
-}
