+++ src/main/java/com/kb04/starroad/Dto/SubscriptionDto.java
@@ -1,36 +1,23 @@
 package com.kb04.starroad.Dto;
 
+import com.kb04.starroad.Dto.product.ProductDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
-
-@Entity
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@SequenceGenerator(name = "subscription_seq", sequenceName="subscription_seq")
-@Table(name = "subscription")
 public class SubscriptionDto {
-    @Id
-    @Column(nullable = false)
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator="subscription_seq")
     private int no;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_no", nullable = false)
     private MemberDto memberNo;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "prod_no", nullable = false)
     private ProductDto prodNo;
 
-    @Column(nullable = false)
     private int period;
 
-    @Column(nullable = false)
     private int price;
 }
