+++ src/main/java/com/kb04/starroad/Dto/policy/PolicyDto.java
@@ -1,0 +1,19 @@
+package com.kb04.starroad.Dto.policy;
+
+import lombok.*;
+
+import javax.persistence.*;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PolicyDto {
+
+    private int no;
+    private String name;
+    private String location;
+    private String explain;
+    private String tag;
+    private String link;
+}
