+++ src/main/java/com/kb04/starroad/Repository/BoardRepository.java
@@ -1,6 +1,6 @@
 package com.kb04.starroad.Repository;
 
-import com.kb04.starroad.Dto.BoardDto;
+import com.kb04.starroad.Dto.board.BoardDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
