+++ src/main/java/com/kb04/starroad/Repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.kb04.starroad.Repository;
 
-import com.kb04.starroad.Dto.ProductDto;
+import com.kb04.starroad.Dto.product.ProductDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
