+++ src/main/java/com/kb04/starroad/Service/BoardService2.java
@@ -1,6 +1,6 @@
 package com.kb04.starroad.Service;
 
-import com.kb04.starroad.Dto.BoardDto;
+import com.kb04.starroad.Dto.board.BoardDto;
 import com.kb04.starroad.Repository.BoardRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -8,7 +8,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.transaction.Transactional;
 import java.io.IOException;
 import java.util.List;
